Moving common Kafka event decoder in commons
