+++ packages/agreement-readmodel-writer/src/index.ts
@@ -1,7 +1,11 @@
 import { EachMessagePayload } from "kafkajs";
-import { consumerConfig, logger } from "pagopa-interop-commons";
+import {
+  consumerConfig,
+  logger,
+  decodeKafkaMessage,
+} from "pagopa-interop-commons";
 import { runConsumer } from "kafka-iam-auth";
-import { decodeKafkaAgreementMessage } from "pagopa-interop-models";
+import { AgreementEvent } from "pagopa-interop-models";
 import { handleMessage } from "./agreementConsumerService.js";
 
 const config = consumerConfig();
@@ -11,7 +15,7 @@ async function processMessage({
   partition,
 }: EachMessagePayload): Promise<void> {
   try {
-    await handleMessage(decodeKafkaAgreementMessage(message));
+    await handleMessage(decodeKafkaMessage(message, AgreementEvent));
 
     logger.info(
       `Read model was updated. Partition number: ${partition}. Offset: ${message.offset}`
