+++ packages/commons/src/index.ts
@@ -8,3 +8,4 @@ export * from "./logging/index.js";
 export * from "./model/apiEndpoint.js";
 export * from "./types/index.js";
 export * from "./fileManager.js";
+export * from "./kafka/kafka.js";
