+++ packages/models/package.json
@@ -25,7 +25,6 @@
     "@protobuf-ts/protoc": "^2.9.1",
     "@protobuf-ts/runtime": "^2.9.1",
     "@types/node": "^20.3.1",
-    "kafkajs": "^2.2.4",
     "ts-pattern": "^5.0.1",
     "zod": "^3.21.4"
   },
