+++ packages/models/src/readModels/events.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import { KafkaMessage } from "kafkajs";
 import { z } from "zod";
 
 export const EventEnvelope = <TEventZodType extends z.ZodType>(
@@ -45,18 +44,3 @@ type MessageT<TEventZodType extends z.ZodType> = ReturnType<
   typeof Message<TEventZodType>
 >;
 export type Message<TEvent> = z.infer<MessageT<z.ZodType<TEvent>>>;
-
-export function decodeKafkaMessage<TEvent extends z.ZodType>(
-  message: KafkaMessage,
-  event: TEvent
-) {
-  const parsed = Message(event).safeParse(message);
-  if (!parsed.success) {
-    throw new Error("Invalid message: " + JSON.stringify(parsed.error));
-  } else if (!parsed.data.value?.after) {
-    throw new Error(
-      "Invalid message: missing value " + JSON.stringify(parsed.data)
-    );
-  }
-  return parsed.data.value.after;
-}
